integrate cadc-data-ops-fits api into test; remove obsolete stuff
