+++ minoc/build.gradle
@@ -40,7 +40,8 @@ dependencies {
     compile 'org.opencadc:cadc-permissions-client:[0.2,)'
     compile 'org.opencadc:cadc-gms:[1.0,)'
     
-    compile 'gov.nasa.gsfc.heasarc:nom-tam-fits:1.15.3-pdowler'
+    compile 'org.opencadc:cadc-data-ops-fits:[0.1,1.0)'
+    compile 'org.opencadc:nom-tam-fits:[1.16.0,2.0)'
     
     testCompile 'junit:junit:[4.0,)'
     testCompile 'org.opencadc:cadc-storage-adapter-fs:[0.6,)'
